safety-macro: define #[safety] macro and add testcases
